web-platform_23_03_dev/requirements.txt
asgiref==3.6.0
cx-Oracle==8.3.0
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-grappelli==3.0.4
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
openpyxl==3.1.2
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
